Moved shared types to root of `types` directory
